refactor: cambiar MileageRecord.reading a Long y notifier a Long
- MileageRecord.reading y MileageEntity.reading ahora Long.
- Elimina el límite Int.MAX_VALUE en el odómetro; acepta cualquier Long no negativo válido.
- AddReadingResult.RequiresConfirmation.previousReading y MileageUiState.pendingConfirmation ahora Long.
- MileageAlertNotifier.onMileageUpdated recibe odometer Long; NoOp y Recording ajustados.
- MaintenanceViewModel convierte la lectura Long a Int para el módulo de mantenimiento existente.
- Tests actualizados a literales L y assertions Long.
- Verificación: assembleDebug + testDebugUnitTest + lintDebug exitosos.

+++ app/src/main/java/com/micarro/domain/alerts/MileageAlertNotifier.kt
@@ -9,9 +9,9 @@ import javax.inject.Inject
  * módulo de alertas; mientras tanto Hilt enlaza [NoOpMileageAlertNotifier].
  */
 interface MileageAlertNotifier {
-    suspend fun onMileageUpdated(vehicleId: String, odometer: Int)
+    suspend fun onMileageUpdated(vehicleId: String, odometer: Long)
 }
 
 class NoOpMileageAlertNotifier @Inject constructor() : MileageAlertNotifier {
-    override suspend fun onMileageUpdated(vehicleId: String, odometer: Int) = Unit
+    override suspend fun onMileageUpdated(vehicleId: String, odometer: Long) = Unit
 }
