+++ app/src/main/java/com/micarro/domain/model/MileageRecord.kt
@@ -6,6 +6,6 @@ data class MileageRecord(
     val id: Long = 0L,
     val vehicleId: Long,
     val date: LocalDate,
-    val reading: Int,
+    val reading: Long,
     val note: String? = null
 )
